Change file to match with the stylecheck script
