+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello I'm July, your helpful assistant
+Hello I'm july, your helpful assistant
 Okie no problem, I've add the given task:
 [T][ ] buy milk
 Okie no problem, I've added a new deadline:
